[store/meta] refactor: move state-machine into standalone file
